+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 
-
 # Changelog
 
 ## RABBIT v1.3.4
 * reset default theshcall = 0.9
 * MagicBase: fix commentlines of vcf
 * MagicImpute: reset default trimfraction = 0.025, default skeletonsize = 100
+* MagicImpute: set tukeyfence = 2 for MagicImpute, and 3 for others (MagicCall and MagicReconstruct). 
+* MagicBase: new function plotmissing
+* MagicBase: default SoftThreshLikeParam to 0.05 for foundererror, offspringerror, and peroffspringerror. 
 
 ## RABBIT v1.3.3
 * magiccall: reset defaults: likeparameters, priorlikeparametes, threshoffspring